requirements.txt
altgraph==0.17.3
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
kiwisolver==1.4.4
macholib==1.16.2
matplotlib==3.6.3
mecab-python3==1.0.6
numpy==1.24.1
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
tabulate==0.9.0
unidic-lite==1.0.8
wordcloud==1.8.2.2
wxPython==4.2.0
